+++ packages/terra-framework-docs/src/terra-dev-site/doc/layout/example/LayoutMenuDisabled.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Layout from 'terra-layout';
-import ContentExample from 'terra-layout/lib/terra-dev-site/doc/common/ContentExample';
-import ToolbarExample from 'terra-layout/lib/terra-dev-site/doc/common/ToolbarExample';
 import classNames from 'classnames/bind';
+import ContentExample from '../common/ContentExample';
+import ToolbarExample from '../common/ToolbarExample';
 import styles from './LayoutDocExampleCommon.module.scss';
 
 const cx = classNames.bind(styles);