+++ packages/terra-framework-docs/src/terra-dev-site/test/layout/layout/LayoutNoMenu.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Layout from '../../../Layout';
+import Layout from 'terra-layout';
 import TestHeader from '../common/TestHeader';
 import TestContent from '../common/TestContent';
 import styles from './TestLayoutCommon.module.scss';